ROUND 1 [STEP 2] 릴라
@릴라
만나뵙게 되서 반갑습니다.
추석은 즐겁게 보내셨는지요?

## 고민했던 점
### 주제
- for in 에서 반복문 배열에 접근하는 변수가 사용하지 않을때는
Immutable value 'column' was never used; consider replacing with '_' or removing it
위의 문구로 경고가 나는데 언더바(_)로 사용하는게 나을까요 아님 이름을 갖는게 나을지요?
- 순서도를 처음 그려봤는데 디테일하게 피드백 부탁드려요.
- 계행안하는 출력함수를 따로 정의해서 사용했습니다.
